1245, 1025

+++ codingtest/baekjun problems/4375.cpp
@@ -4,8 +4,8 @@ using namespace std;
 int main(void){
     int n;
     while(cin >> n){
-        long long v = 1;
-        short count = 1;
+        int v = 1;
+        int count = 1;
 
         while(v%n!=0){
             v = ((v*10)+1)%n;
